get-video/requirements.txt
fastapi
uvicorn[standard]
google-api-python-client
google-auth-oauthlib
google-auth
anthropic
requests
pydantic>=2.0
pydantic-settings
python-dotenv
